Location handling and renaming
Introduce a config file to describe the bounds for locations and their
names. When an activity's start or end location falls within those
boundaries then the activity is renamed and tagged as "commute".

I thought about testing the `Config` object directly but it feels more
thorough to test all the way through the handler. Especially as it's not
feasible to do end-to-end integration testing.