Set gear ID if present in config
So that you can associate all commutes with a single bike.

I considered making this a property of the location but it's not a
feature that I need right now and it means figuring out what to do if an
activity starts and ends with locations that have different bikes.